Modified Styles, Renamed Search Added Tui
