+++ client/src/components/Main.js
@@ -5,100 +5,11 @@ import CssBaseline from "@material-ui/core/CssBaseline";
 import Paper from "@material-ui/core/Paper";
 import Typography from "@material-ui/core/Typography";
 import Grid from "@material-ui/core/Grid";
-
 import Topbar from "./Topbar";
+import styles from "./styles/Styles"
 
-const backgroundShape = require("../images/shape.svg");
+const tuiImage = require("../images/team-tui.jpg");
 
-const styles = theme => ({
-  root: {
-    flexGrow: 1,
-    backgroundColor: theme.palette.grey["100"],
-    overflow: "hidden",
-    background: `url(${backgroundShape}) no-repeat`,
-    backgroundSize: "cov er",
-    backgroundPosition: "0 400px",
-    paddingBottom: 200
-  },
-  grid: {
-    width: 1200,
-    margin: `0 ${theme.spacing(2)}px`,
-    [theme.breakpoints.down("sm")]: {
-      width: "calc(100% - 20px)"
-    }
-  },
-  loadingState: {
-    opacity: 0.05
-  },
-  paper: {
-    padding: theme.spacing(3),
-    margin: theme.spacing(2),
-    textAlign: "left",
-    color: theme.palette.text.secondary
-  },
-  rangeLabel: {
-    display: "flex",
-    justifyContent: "space-between",
-    paddingTop: theme.spacing(2)
-  },
-  topBar: {
-    display: "flex",
-    justifyContent: "space-between",
-    alignItems: "center"
-  },
-  outlinedButtom: {
-    textTransform: "uppercase",
-    margin: theme.spacing(1)
-  },
-  actionButtom: {
-    textTransform: "uppercase",
-    margin: theme.spacing(1),
-    width: 152,
-    height: 36
-  },
-  blockCenter: {
-    padding: theme.spacing(2),
-    textAlign: "center"
-  },
-  block: {
-    padding: theme.spacing(2)
-  },
-  loanAvatar: {
-    display: "inline-block",
-    verticalAlign: "center",
-    width: 16,
-    height: 16,
-    marginRight: 10,
-    marginBottom: -2,
-    color: theme.palette.primary.contrastText,
-    backgroundColor: theme.palette.primary.main
-  },
-  interestAvatar: {
-    display: "inline-block",
-    verticalAlign: "center",
-    width: 16,
-    height: 16,
-    marginRight: 10,
-    marginBottom: -2,
-    color: theme.palette.primary.contrastText,
-    backgroundColor: theme.palette.primary.light
-  },
-  inlining: {
-    display: "inline-block",
-    marginRight: 10
-  },
-  buttonBar: {
-    display: "flex"
-  },
-  noBorder: {
-    borderBottomStyle: "hidden"
-  },
-  mainBadge: {
-    textAlign: "center",
-    marginTop: theme.spacing(4),
-    marginBottom: theme.spacing(4)
-  }
-});
 
 class Main extends Component {
   state = {
@@ -156,6 +67,15 @@ class Main extends Component {
                           Welcome to the SEER Research Searching Tool
                           This is our Home page
                         </Typography>
+                      </div>
+                    </div>
+                  </Paper>
+                </Grid>
+                <Grid item xs={12}>
+                  <Paper className={classes.paper}>
+                    <div>
+                      <div className={classes.box}>
+                        <img className={classes.img} src={tuiImage} alt={"Tui's are lit"} />
                       </div>
                     </div>
                   </Paper>
