+++ client/src/routes.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Route, HashRouter, Switch } from 'react-router-dom'
 import Main from './components/Main'
-import Dashboard from './components/Dashboard'
+import Search from './components/Search'
 import ScrollToTop from './components/ScrollTop'
 
 export default props => (
@@ -9,7 +9,7 @@ export default props => (
         <ScrollToTop>
             <Switch>
                 <Route exact path='/' component={Main} />
-                <Route exact path='/dashboard' component={Dashboard} />
+                <Route exact path='/search' component={Search} />
             </Switch>
         </ScrollToTop>
     </HashRouter>
